refactor: #332 StationUseCase 프로토콜 분리
